added (untested) support for big endian machines

+++ Renderer/GLRenderer.cpp
@@ -618,7 +618,7 @@ bool GLRenderer::Execute3DFrame(ERenderArea eREnderArea) {
 
   // if there is something left in the TODO list
   if (m_vCurrentBrickList.size() > m_iBricksRenderedInThisSubFrame) {
-    m_pMasterController->DebugOut()->Message("GLRenderer::Execute3DFrame","%i bricks left to render", m_vCurrentBrickList.size()-m_iBricksRenderedInThisSubFrame);
+    m_pMasterController->DebugOut()->Message("GLRenderer::Execute3DFrame","%i bricks left to render", int(UINT64(m_vCurrentBrickList.size())-m_iBricksRenderedInThisSubFrame));
 
     // setup shaders vars
     SetDataDepShaderVars(); 
@@ -970,7 +970,7 @@ void GLRenderer::Render3DView() {
   timeStart = timeProbe = clock();
 
   while (m_vCurrentBrickList.size() > m_iBricksRenderedInThisSubFrame && float(timeProbe-timeStart)*1000.0f/float(CLOCKS_PER_SEC) < m_iTimeSliceMSecs) {
-    m_pMasterController->DebugOut()->Message("GLRenderer::Render3DView","  Brick %i of %i", m_iBricksRenderedInThisSubFrame+1,int(m_vCurrentBrickList.size()));
+    m_pMasterController->DebugOut()->Message("GLRenderer::Render3DView","  Brick %i of %i", int(m_iBricksRenderedInThisSubFrame+1),int(m_vCurrentBrickList.size()));
 
     // convert 3D variables to the more general ND scheme used in the memory manager, e.i. convert 3-vectors to stl vectors
     vector<UINT64> vLOD; vLOD.push_back(m_iCurrentLOD);
